Move graph fixtures to test_utils.graphs
